chore: change the name of the folders
- app to presentation
- move route under the presentation
- move util folder to root
