+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:kakao_flutter_sdk_common/kakao_flutter_sdk_common.dart';
 
-import 'app/theme/theme.dart';
-import 'app/utils/navigator.dart';
+import 'presentation/routes/route.dart';
+import 'presentation/theme/theme.dart';
 
 void main() async {
   await dotenv.load(fileName: ".env");
@@ -11,7 +11,6 @@ void main() async {
   // runApp() 호출 전 Flutter SDK 초기화
   KakaoSdk.init(
     nativeAppKey: dotenv.env['NATIVE_APP_KEY'],
-    javaScriptAppKey: dotenv.env['JAVASCRIPT_APP_KEY'],
   );
 
   runApp(const MyApp());
@@ -26,7 +25,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp.router(
       title: 'Flutter Demo',
       theme: lightTheme,
-      routerConfig: navigator,
+      routerConfig: router,
     );
   }
 }
