+++ lib/presentation/pages/sign_up/views/styles.dart
