+++ lib/presentation/pages/sign_up_details/views/sign_up_details_0123.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
+import '../../../../utils/constants.dart';
+import '../../../routes/route_path.dart';
 import '../../../theme/colors.dart';
 import '../../../theme/styles.dart';
-import '../../../utils/constants.dart';
 import '../../../widgets/app_bar.dart';
 import '../../../widgets/snack_bar.dart';
 import '../../../widgets/text.dart';
@@ -11,25 +12,27 @@ import '../widgets/buttons.dart';
 import '../widgets/steppers.dart';
 import 'styles.dart';
 
-class SignUpDetail0126 extends StatefulWidget {
-  const SignUpDetail0126({super.key});
+class SignUpDetail0123 extends StatefulWidget {
+  const SignUpDetail0123({super.key});
 
   @override
-  State<SignUpDetail0126> createState() => _SignUpDetail0126State();
+  State<SignUpDetail0123> createState() => _SignUpDetail0121State();
 }
 
-class _SignUpDetail0126State extends State<SignUpDetail0126> {
+class _SignUpDetail0121State extends State<SignUpDetail0123> {
   bool _isButtonDisabled = true;
   final List<int> selectedItems = [];
 
-  final List<String> challengeList =
-      ChallengeList.values.map((element) => element.option).toList();
+  final timeLabelList = Time.values.map((element) => element.label).toList();
+  final timeHoursList = Time.values.map((element) => element.hours).toList();
 
   void _selectTiles(int index) {
     setState(() {
+      // 빈 배열 인경우 (아무 것도 선택 안 했을 때)
       if (selectedItems.isEmpty) {
         _isButtonDisabled = false;
         selectedItems.add(index);
+        // 요소가 하나 있는 경우 (한 개 선택 했을 때)
       } else if (selectedItems.length == 1) {
         if (selectedItems.contains(index)) {
           selectedItems.remove(index);
@@ -37,6 +40,7 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
         } else {
           selectedItems.add(index);
         }
+        // 요소가 두개 있는 경우 (모두 선택 했을 때)
       } else if (selectedItems.length == 2) {
         if (selectedItems.contains(index)) {
           selectedItems.remove(index);
@@ -50,7 +54,7 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
 
   void _navigateToNextPage() {
     if (selectedItems.isNotEmpty) {
-      context.push('/sign_up/detail/0127');
+      context.push('${RouterPath.signUp}/detail/0124');
     }
   }
 
@@ -63,7 +67,7 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
         children: [
           const SignUpDetailsStepper(
             currentStep: 2,
-            stage: '성향선택(4/4)',
+            stage: '성향선택(1/4)',
           ),
           Padding(
             padding: const EdgeInsets.only(left: 20.0, top: 10.0, right: 20.0),
@@ -71,8 +75,8 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 buildTitleText(
-                  '아래 항목 중\n도전하고 싶은 것은 무엇인가요?',
-                  '최대 2개까지 선택할 수 있어요.',
+                  '주로 작업하는 시간대는\n어떻게 되시나요?',
+                  '비슷한 성향의 파티원을 추천해 드려요. (최대 2개)',
                 ),
                 _buildListView(),
               ],
@@ -87,7 +91,7 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
           ),
           buildSkipButton(
             context,
-            '/sign_up/detail/0125',
+            '${RouterPath.signUp}/detail/0124',
           ),
         ],
       ),
@@ -104,13 +108,14 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
         separatorBuilder: (BuildContext context, int index) {
           return const SizedBox(height: 8.0);
         },
-        itemCount: challengeList.length,
+        itemCount: timeLabelList.length,
       ),
     );
   }
 
   Widget _buildListTiles(int index) {
-    final personality = challengeList[index];
+    final timeLabel = timeLabelList[index];
+    final timeHours = timeHoursList[index];
     final bool isSelected = selectedItems.contains(index);
 
     return Material(
@@ -120,10 +125,11 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
           : SignUpDetailsListStyle.unselectedBorder,
       child: ListTile(
         contentPadding: const EdgeInsets.only(left: 20),
-        title: Text(personality),
+        title: Text('$timeLabel $timeHours'),
         titleTextStyle: isSelected
             ? SignUpDetailsListStyle.selectedText
             : SignUpDetailsListStyle.unselectedText,
+        // 클릭한 타일 (타일 5개 중에 무엇이 selected 됐는지 확인)
         selected: isSelected,
         onTap: () {
           setState(() {
@@ -139,9 +145,7 @@ class _SignUpDetail0126State extends State<SignUpDetail0126> {
         selectedTileColor: AppColors.primaryLightColors.shade300,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(16.0),
-          side: BorderSide(
-            color: AppColors.greyColors.shade200,
-          ),
+          side: BorderSide(color: AppColors.greyColors.shade200),
         ),
       ),
     );
