+++ lib/presentation/pages/sign_up_details/views/styles.dart
