+++ lib/presentation/pages/sign_up_details/widgets/steppers.dart
