+++ lib/presentation/routes/route_path.dart
@@ -1,0 +1,8 @@
+class RouterPath {
+  static const String main = '/main';
+
+  static const String signIn = '/sign_in';
+
+  static const String signUp = '/sign_up';
+  static const String signUpDetail = 'sign_up/detail';
+}
